+++ tests/integration-test-config.php
@@ -238,6 +238,7 @@ return [
 		'SecurityCheck-CUSTOM2',
 		'SecurityCheck-OTHER',
 		'SecurityCheck-LikelyFalsePositive',
+		'SecurityCheckDebugTaintedness',
 		// These are to make testing easier
 		'PhanDebugAnnotation',
 		// Uncomment when needed, cannot be left in place because it might be emitted